Enhance docker-compose and configuration files

+++ init.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
+if ! command -v docker &> /dev/null; then
+    echo "Docker could not be found. Please install Docker and try again."
+    exit 1
+fi
+
 bash ./create_network.sh
 
-cd ./docker && cp .env.example .env && cd ./..
+if [ ! -f .env ]; then
+    cd ./docker && cp .env.example .env && cd ./..
+fi
+
+docker compose up -d
